+++ Tests/Runtime/CommandDispatchTests/CommandDispatchTests.cs
@@ -31,7 +31,8 @@ namespace Yarn.Unity.Tests
 
         public void Setup()
         {
-            if (Directory.Exists(TestFilesDirectoryPath) == false) {
+            if (Directory.Exists(TestFilesDirectoryPath) == false)
+            {
                 UnityEditor.AssetDatabase.CreateFolder("Assets", TestFolderName);
                 UnityEditor.AssetDatabase.CopyAsset(TestScriptPathSource, TestScriptPathInProject);
             }
@@ -57,7 +58,8 @@ namespace Yarn.Unity.Tests
         }
 
         [Test]
-        public void CommandDispatch_Passes() {
+        public void CommandDispatch_Passes()
+        {
             var dialogueRunnerGO = new GameObject("Dialogue Runner");
             var dialogueRunner = dialogueRunnerGO.AddComponent<DialogueRunner>();
 
@@ -90,6 +92,53 @@ namespace Yarn.Unity.Tests
             {
                 Assert.True(dialogueRunner.Dialogue.Library.FunctionExists(expectedFunctionName), "expected function {0} to be registered", expectedFunctionName);
             }
+        }
+
+        [Test]
+        public void DirectRegistrationCommands_CanHaveParamsArray()
+        {
+            void TestCommand(params string[] parameters)
+            {
+                Debug.Log(string.Join(";", parameters));
+            }
+
+            void InvalidArrayType(params int[] ints)
+            {
+                Assert.Fail("This method should not be called");
+            }
+
+            void InvalidArrayPosition(string[] strings, bool boolean)
+            {
+                Assert.Fail("This method should not be called");
+            }
+
+            var dialogueRunnerGO = new GameObject("Dialogue Runner");
+            var dialogueRunner = dialogueRunnerGO.AddComponent<DialogueRunner>();
+
+            dialogueRunner.AddCommandHandler<string[]>("test_command", TestCommand);
+
+            LogAssert.Expect(LogType.Log, "1;2;3;4");
+
+            var dispatchResult = dialogueRunner.CommandDispatcher.DispatchCommand("test_command 1 2 3 4", out Coroutine result);
+
+            Assert.IsTrue(dispatchResult.IsSuccess);
+            Assert.AreEqual(dispatchResult.Status, CommandDispatchResult.StatusType.SucceededSync);
+            Assert.IsNull(result);
+
+            var invalidTypeException = Assert.Throws<System.ArgumentException>(() =>
+            {
+                dialogueRunner.AddCommandHandler<int[]>("invalid_array_type", InvalidArrayType);
+            }, "command handler array parameters must be arrays of strings");
+            Assert.True(invalidTypeException.Message.Contains("array parameters are required to be string arrays"));
+
+
+            var notLastParameterException = Assert.Throws<System.ArgumentException>(() =>
+            {
+                dialogueRunner.AddCommandHandler<string[], bool>("invalid_array_type", InvalidArrayPosition);
+            }, "command handler array parameters must be the last parameter");
+            Assert.True(notLastParameterException.Message.Contains("array parameters are required to be last"));
+
+
         }
     }
 }